@@ -7,7 +7,6 @@
     "build": "cross-env NODE_ENV=production webpack --mode=production --progress --hide-modules"
   },
   "dependencies": {
-    "moment": "2.24.0",
     "vue": "^2.5.22"
   },
   "devDependencies": {
@@ -21,23 +20,24 @@
     "babel-preset-env": "^1.7.0",
     "babel-preset-stage-3": "^6.24.1",
     "cross-env": "^7.0.2",
-    "css-loader": "^3.5.2",
+    "css-loader": "^3.6.0",
     "eslint": "^5.8.0",
     "eslint-plugin-vue": "^5.0.0",
     "expect": "^26.1.0",
+    "eslint-plugin-vue": "^6.2.2",
     "file-loader": "^6.0.0",
     "jsdom": "^16.2.2",
     "jsdom-global": "^3.0.2",
-    "mocha": "^7.1.1",
+    "mocha": "^8.0.1",
     "mocha-webpack": "^1.1.0",
-    "node-sass": "^4.13.1",
+    "node-sass": "^4.14.1",
     "regenerator-runtime": "^0.13.5",
     "sass-loader": "^8.0.2",
     "uglifyjs-webpack-plugin": "^2.2.0",
     "vue-loader": "^15.9.1",
     "vue-loader-plugin": "^1.3.0",
     "vue-template-compiler": "^2.6.11",
-    "webpack": "^4.42.1",
+    "webpack": "^4.43.0",
     "webpack-cli": "^3.3.11",
     "webpack-dev-server": "^3.10.3",
     "webpack-node-externals": "^1.7.2"
