@@ -1,10 +1,8 @@
-import Vue from "vue";
-import App from "./App.vue";
-
-window.moment = require("moment");
+import Vue from 'vue';
+import App from './App.vue';
 
 Vue.config.productionTip = false;
 
 new Vue({
-  render: h => h(App)
-}).$mount("#app");
+    render: (h) => h(App),
+}).$mount('#app');
