@@ -1,18 +1,18 @@
 export const arrowKeys = [
-  {
-    code: 37,
-    direction: "left"
-  },
-  {
-    code: 38,
-    direction: "up"
-  },
-  {
-    code: 39,
-    direction: "right"
-  },
-  {
-    code: 40,
-    direction: "down"
-  }
+    {
+        code: 37,
+        direction: 'left',
+    },
+    {
+        code: 38,
+        direction: 'up',
+    },
+    {
+        code: 39,
+        direction: 'right',
+    },
+    {
+        code: 40,
+        direction: 'down',
+    },
 ];
